+++ essence-of-live-coding/src/LiveCoding.hs
@@ -24,6 +24,15 @@ import LiveCoding.Exceptions as X
 import LiveCoding.Exceptions.Finite as X
 import LiveCoding.Forever as X
 import LiveCoding.Handle as X
+import LiveCoding.HandlingState as X
+    ( HandlingStateT,
+      HandlingState(..),
+      Handling(..),
+      isRegistered,
+      runHandlingStateT,
+      runHandlingStateC,
+      runHandlingState,
+    )
 import LiveCoding.Handle.Examples as X
 import LiveCoding.LiveProgram as X
 import LiveCoding.LiveProgram.HotCodeSwap as X
