+++ essence-of-live-coding/src/LiveCoding/Handle.hs
@@ -1,47 +1,27 @@
 {-# LANGUAGE Arrows #-}
-{-# LANGUAGE DeriveDataTypeable #-}
-{-# LANGUAGE ExistentialQuantification #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
-{-# LANGUAGE GADTs #-}
 {-# LANGUAGE RecordWildCards #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 
-module LiveCoding.Handle
-  ( Handle (..)
-  , handling
-  , HandlingState (..)
-  , HandlingStateT
-  , isRegistered
-  , runHandlingState
-  , runHandlingStateC
-  , runHandlingStateT
-  )
-  where
+module LiveCoding.Handle where
 
 -- base
-import Control.Arrow (returnA, arr, (>>>))
 import Data.Data
-
--- containers
-import Data.IntMap
-import qualified Data.IntMap as IntMap
 
 -- transformers
 import Control.Monad.Trans.Class (MonadTrans(lift))
-import Control.Monad.Trans.State.Strict
 
 -- essence-of-live-coding
 import LiveCoding.Cell
-import LiveCoding.Cell.Monad
-import LiveCoding.Cell.Monad.Trans
-import LiveCoding.LiveProgram
-import LiveCoding.LiveProgram.Monad.Trans
+import LiveCoding.HandlingState
 
 {- | Container for unserialisable values,
 such as 'IORef's, threads, 'MVar's, pointers, and device handles.
 
-In a 'Handle', you can store a mechanism to create and destroy a value that survives live coding even if does not have a 'Data' instance.
+In a 'Handle', you can store a mechanism to create and destroy a value
+that survives reloads occuring during live coding
+even if does not have a 'Data' instance.
 Using the function 'handling', you can create a cell that will
 automatically initialise your value,
 and register it in the 'HandlingStateT' monad transformer,
@@ -71,83 +51,6 @@ combineHandles handle1 handle2 = Handle
   , destroy = \(h1, h2) -> destroy handle2 h2 *> destroy handle1 h1
   }
 
-data Handling h where
-  Handling
-    :: { id     :: Key
-       , handle :: h
-       }
-    -> Handling h
-  Uninitialized :: Handling h
-
-type Destructors m = IntMap (Destructor m)
-
--- | Hold a map of registered handle keys and destructors
-data HandlingState m = HandlingState
-  { nHandles    :: Key
-  , destructors :: Destructors m
-  }
-  deriving Data
-
--- | In this monad, handles can be registered,
---   and their destructors automatically executed.
---   It is basically a monad in which handles are automatically garbage collected.
-type HandlingStateT m = StateT (HandlingState m) m
-
-initHandlingState :: HandlingState m
-initHandlingState = HandlingState
-  { nHandles = 0
-  , destructors = IntMap.empty
-  }
-
--- | Handle the 'HandlingStateT' effect _without_ garbage collection.
---   Apply this to your main loop after calling 'foreground'.
---   Since there is no garbage collection, don't use this function for live coding.
-runHandlingStateT
-  :: Monad m
-  => HandlingStateT m a
-  -> m a
-runHandlingStateT = flip evalStateT initHandlingState
-
-{- | Apply this to your main live cell before passing it to the runtime.
-
-On the first tick, it initialises the 'HandlingState' at "no handles".
-
-On every step, it does:
-
-1. Unregister all handles
-2. Register currently present handles
-3. Destroy all still unregistered handles
-   (i.e. those that were removed in the last tick)
--}
-runHandlingStateC
-  :: forall m a b .
-     (Monad m, Typeable m)
-  => Cell (HandlingStateT m) a b
-  -> Cell                 m  a b
-runHandlingStateC cell = flip runStateC_ initHandlingState
-  $ hoistCellOutput garbageCollected cell
-
--- | Like 'runHandlingStateC', but for whole live programs.
-runHandlingState
-  :: (Monad m, Typeable m)
-  => LiveProgram (HandlingStateT m)
-  -> LiveProgram                 m
-runHandlingState LiveProgram { .. } = flip runStateL initHandlingState LiveProgram
-  { liveStep = garbageCollected . liveStep
-  , ..
-  }
-
-garbageCollected
-  :: Monad m
-  => HandlingStateT m a
-  -> HandlingStateT m a
-garbageCollected action = unregisterAll >> action <* destroyUnregistered
-
-data Destructor m = Destructor
-  { isRegistered :: Bool
-  , action       :: m ()
-  }
-
 {- | Hide a handle in a cell,
 taking care of initialisation and destruction.
 
@@ -174,85 +77,6 @@ handling handleImpl@Handle { .. } = Cell
         return (handle, state)
       Uninitialized -> do
         handle <- lift create
-        id <- register (destroy handle) handle
+        key <- register (destroy handle) handle
         return (handle, Handling { .. })
   }
-
-register
-  :: Monad m
-  => m () -- ^ Destructor
-  -> h -- ^ Handle value
-  -> HandlingStateT m Key
-register destructor handle = do
-  HandlingState { .. } <- get
-  let id = nHandles + 1
-  put HandlingState
-    { nHandles = id
-    , destructors = insertDestructor destructor id handle destructors
-    }
-  return id
-
-reregister
-  :: Monad m
-  => m ()
-  -> Handling h -- FIXME Unsafe since this might be Uninitialized
-  -> HandlingStateT m ()
-reregister action Handling { .. } = do
-  HandlingState { .. } <- get
-  put HandlingState { destructors = insertDestructor action id handle destructors, .. }
-
-insertDestructor
-  :: m ()
-  -> Key
-  -> h
-  -> Destructors m
-  -> Destructors m
-insertDestructor action id handle destructors =
-  let destructor = Destructor { isRegistered = True, .. }
-  in  insert id destructor destructors
-
-unregisterAll
-  :: Monad m
-  => HandlingStateT m ()
-unregisterAll = do
-  HandlingState { .. } <- get
-  let newDestructors = IntMap.map (\destructor -> destructor { isRegistered = False }) destructors
-  put HandlingState { destructors = newDestructors, .. }
-
-destroyUnregistered
-  :: Monad m
-  => HandlingStateT m ()
-destroyUnregistered = do
-  HandlingState { .. } <- get
-  let
-      (registered, unregistered) = partition isRegistered destructors
-  traverse (lift . action) unregistered
-  put HandlingState { destructors = registered, .. }
-
--- * 'Data' instances
-
-dataTypeHandling :: DataType
-dataTypeHandling = mkDataType "Handling" [handlingConstr, uninitializedConstr]
-
-handlingConstr :: Constr
-handlingConstr = mkConstr dataTypeHandling "Handling" [] Prefix
-
-uninitializedConstr :: Constr
-uninitializedConstr = mkConstr dataTypeHandling "Uninitialized" [] Prefix
-
-instance (Typeable h) => Data (Handling h) where
-  dataTypeOf _ = dataTypeHandling
-  toConstr Handling { .. } = handlingConstr
-  toConstr Uninitialized = uninitializedConstr
-  gunfold _cons nil constructor = nil Uninitialized
-
-dataTypeDestructor :: DataType
-dataTypeDestructor = mkDataType "Destructor" [ destructorConstr ]
-
-destructorConstr :: Constr
-destructorConstr = mkConstr dataTypeDestructor "Destructor" [] Prefix
-
-instance Typeable m => Data (Destructor m) where
-  dataTypeOf _ = dataTypeDestructor
-  toConstr Destructor { .. } = destructorConstr
-  gunfold _ _ = error "Destructor.gunfold"
